requirements.txt
google-api-python-client==2.108.0
pandas==2.1.3
numpy==1.26.2
streamlit==1.28.2
plotly==5.18.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
fastapi==0.104.1
uvicorn==0.24.0
python-dateutil==2.8.2
